cli: Fix multisig parsing (#72)
* cli: Fix multisig parsing

#### Problem

As pointed out in #58, the CLI currently fails when just running
`create-account <MINT_ADDRESS>`, because a multisig argument is expected
during all commands. Clap v3's version of `values_of` gives an error if
the arg is not configured in the command.

This was probably missed during the port over to clap v3.

#### Summary of changes

Do the same thing as elsewhere, and change `values_of` to
`try_get_many(...).ok().flatten()`.

This is the last place using `values_of` in the CLI that could fail.

* Integrate review feedback