CI: Update tag name to match the repo
#### Problem

The tag names from SPL contained nearly the full name of the package,
which made sense because there were many packages. The program-specific
repos, however, take a different format, and only keep the basename of
the directory. Also, they use `@` to separate the name from the version.

#### Summary of changes

Use the basename and separate with `@`!