program: Refactor multisig support to not require stubs
#### Problem

To lighten the dependencies on token-2022, we need to switch from using
`solana-program` to the component crates, but the component crates don't
properly support the program-stubs for CPI, ie.
`solana_program::program::invoke` uses the stubs, but
`solana_cpi::invoke` does not. This means that tests using the stubs
fail after the transition.

#### Summary of changes

Refactor the code to return the instruction and account infos, so that
it can be directly tested without the use of `invoke_signed`.