program: Remove solana-program / sdk (#264)
* program: Remove solana-program / sdk

#### Problem

The token-2022 program still uses solana-program and solana-sdk, which
are heavy dependencies.

#### Summary of changes

The component crates exist to do everything needed in token-2022, so we
can remove solana-program usage entirely.

NOTE: since some spl-token-2022's dependencies still use solana-program,
there is no huge gain just yet. That will happen when we upgrade the
upstream dependencies and then upgrade token-2022 to use those.

Also NOTE: this is a breaking change since we've removed the re-export
of the sdk types. Considering we're now respecting semver, this
shouldn't be a problem, but we did add a version of re-exports in
https://github.com/solana-program/token/pull/26, so maybe we should do
the same thing here. Let me know!

* Fixup tests to run program in bpf mode