program: Relax extension type check to the base number (#546)
#### Problem

As noted in #545, `get_extension_indices` fails if an older version of
the library is trying to parse a new extension type, but the iteration
could otherwise succeed.

#### Summary of changes

Compare the extension type through its underlying byte value instead of
as an enum. Note that this removes an additional check that the
extension type being searched for is valid for the account, ie that
we're not trying to find a mint extension in a token account. That
check, however, was overly defensive, so it likely ins't needed.