deps: Bump sdk in interface and its dependents
#### Problem

The sdk v3 crates are all out, and most of the SPL interface crates have
been updated to use it too, but token-2022 is still on sdk v2.

#### Summary of changes

Bump the interface, proof-extraction, and proof-generation to the newest
SDK crates, and have elgamal-registry / program / rust client all use
the crate versions from crates.io.

At the same time, update the interface to use the spl-token-interface,
and update the error type to use the new traits.