Move repo to work with a Makefile
#### Problem

The program repos are getting a little out of hand, with many different
methods of managing CI, scripts, etc etc.

#### Summary of changes

Token-2022 is the most complicated repo, with many different clients,
multiple programs, and sub-libraries, so if we can get it to work here,
it should be usable in all the other program repos.

* Create a Makefile, similar to the system repo, to handle all jobs
* Convert publish scripts to bash, similar to the system repo again
* Remove js scripts that aren't related to Codama generation
* Update the main CI and publish jobs to use Makefile commands
* Use matrices in CI to simplify job construction

The biggest issue is that Makefile targets can't contain slashes, so we
need to have some mapping between something like `clients-rust-legacy`
to `clients/rust-legacy`. Rather than specify that mapping by hand, I
opted to assume that the first word in a target is a subdirectory. This
means we have to change `confidential-transfer` to `confidential` in the
repo.