deps: Upgrade the rest of the repo to sdk v3 (#797)
* deps: Upgrade the rest of the repo to sdk v3

#### Problem

The program / client / cli crates are still on SDK v2.

#### Summary of changes

Mostly straightforward, but upgrade everything to SDK v3! Intra-repo
dependencies are restored for the most part, and breaking changes are
dealt with as needed.

The most noteworthy change is the addition of a `TryFrom<u32>`
implementation for `TokenError`, which allows us to print the error
appropriately.

* rust-legacy: Update to newest crates, use interfaces

* cli: Update to newest crates, use interfaces

* Update to CLI v3.0.0

* Update lockfile

* Take into account auto-demotion in transfer-hook

* Use existing num_traits implementation