confidential: Bump solana-curve25519 to v3
#### Problem

The confidential crates are still using solana-curve25519 v2 because
that crate is still part of Agave, and v3 wasn't available when the
interface crates were bumped. Thankfully, it didn't cause an issue at
the time.

#### Summary of changes

Get things consistent, and bump the crate to v3 everywhere. In the
future, we should consider moving the code down to the SDK or finding
some other solution to avoid circular dependencies.