repo: Add `debug = "line-tables-only"` to dev profile to reduce build size (#823)
* repo: Add `debug = false` to dev profile to reduce build size

#### Problem

Some of the CI jobs time out or fail because there's too much disk space
used by the build.

#### Summary of changes

Take a hammer to the problem, and just don't emit debug info. On my
machine, this reduces the build size from 21GB to 4.1GB for the rust
legacy build, which always runs into issues.

* Actually, use "line-tables-only"